Include ("/PATH/file.js") is now working fine
REVIEW: With this new include method, debugger can show the the list of included files, and it's content, and no more only a merge of all the files. It's make easy the scripts' debuging.
This 'new' include method can be used together with the 'old' one and also control cycles and multiple incudes (include files once).
Thanks to Peter and QtScript-binding guys: http://mail.kde.org/pipermail/qtscript-bindings/2008-August/000012.html