New DS Plugin: Rooted tree
This is plugin developed to be used in a DS class.

With this plugin, is possible to write scritps  to work with
rooted trees and get a visual feedback about the memory structure
after the end of script run. Is possible to visualize the nodes as
simple nodes or as complete struture where all the pointers are showed.

The nodes can contain 1 to n children and each node can have a diferent
number.