Make Data an ancestor of Group.
This allows handling of groups similar to data elements. Especially,
handling of pointers and positions can be done by a uniform interface.
The reasoning behind this hierarchy is that we can now specify subgraphs
and handle them as if they would have an internal logic/structure.
The data structure plugin then has to take care (in case groups are
supported) to correctly expand/collapse groups for display.