Fix the dynamic property name change in properties dialog.
There was a problem in dynamic problems with some languages (portuguese
for instance), the user cannot change the property name.

Also, the generated propreties cannot be used in scripts, since they
contains ' ' (white spaces).

This patch fix it by replacing the non letters, numbers ou '_' of the
generated names of properties by '_' and rejecting the property name
from user that not start by a letter and contains only letter,numbers
 and '_'