Fix distance computation for unconnected nodes.
If a path does not exist between any two vertices in a graph, the
distance between those vertices is defined to be infinity. The
GraphStructure::distances method was returning zero in this situation
and so this was a bug. A fix for this bug was implemented, a test case
was added to verify these changes, and related documentation was
updated.

Introducing two new strings, as requested at i18n list.

REVIEW: 110822
CCMAIL: kde-i18n-doc@kde.org