Start work on GraphTheory library.
This library shall eventually replace the RocsCore library at some point
in the future. It will be designed from scratch with the lessons learned
from all the major refactorings of RocsCore. By starting from scratch we
will be able to design a leaner and more Rocs-tailored library than
working on the old and cluttered codebase that saw too many well-meant but
not fully completed refactorings and half-done features.