Break update loop.
When moving a node over the left/top border, the node's position change
triggers a moving of the origin coordinate. By changing the origin, also
the node has to be relocated, setting its position and hence again enters
this update loop.
For now we lock the current update state in a private variable. This is
not the most elegant solution, but will vanish after the next refactoring
of the scene class.