Rewrite project handling.
Due to historic reasons, project handling was implemented only half-hearted.
This rewrite of the project class changes this way: Now projects build the center
of a user's session in Rocs, instead of an arbitrary combination of scrips and graphs.
By this, is massively simplifies the file handling and as a plus is also nicely
unit-tested.

DIGEST: Rewrite project support