Fix crash with script engine's garbage collector.
If object ownership is controlled by script engine, the garbage
collector might start too early to delete objects, which leads to
crashes when we call "object->metaObject()". This patch changes all
object ownerships to the Qt ownership model and guarantees that they
are deleted at the end of a script run.

Note that this leads to a big data consumption if during a single run
a lot of edge/node objects are created and deleted and has to be improved.

Side note: fixing this was only possible by the good unit test coverage :)