Implement moving of multiple nodes at once.
The move action is implemented by using a special moving state in the
DSM. At start every selected node binds the movement delta to its
position, and at end this binding is replaced by the exact values.
The scene action also provides a new parameter that tells if the current
press-action was started at a node or not.