Revert "Bump CMake dependency to 3.0.0."
The bump was done preventively and I just verified that 2.8 is
sufficient enough to build the source. So, drop the requirement down again.
I'd suggest bumping when someone complains that it doesn't build anymore
or simply bumping by the end of April is fine as well for me as then I
don't need cmake 2.8 builds anymore ;)

Revert as discussed on IRC.

This reverts commit c9da7aa70ecfc6852dbacfe27a23c8d765cd2d51.

CCMAIL: cordlandwehr@kde.org